tests/wpt/web-platform-tests/ambient-light/AmbientLightSensor_stop.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>AmbientLightSensor Test: stop()</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://www.w3.org/TR/ambient-light/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

var sensor, stop_return;
setup(function() {
  sensor = new AmbientLightSensor();
  sensor.start();
  stop_return = sensor.stop();
});

test(function() {
  assert_equals(String(sensor.reading), "null");
}, "the sensor.reading is null after executing stop() method");

test(function() {
  assert_throws("InvalidStateError", function() { sensor.stop(); }, "stop() twice");
}, "throw an InvalidStateError exception when state is either idle or errored");

test(function() {
  assert_equals(stop_return, undefined);
}, "the sensor.stop() returns undefined");

</script>
